requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Data handling
pandas>=2.2.0
beautifulsoup4==4.12.2
lxml>=5.0.0

# Database
# sqlite3 is built-in to Python
psycopg2-binary>=2.9.9  # PostgreSQL adapter (use >=2.9.9 for Python 3.13 support)
sqlalchemy>=2.0.23  # Database abstraction

# AI/ML
anthropic>=0.40.0
sentence-transformers>=2.2.0  # Semantic similarity for job filtering

# Email handling
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# CV parsing dependencies
pdfplumber==0.10.3      # PDF parsing (better than PyPDF2)
python-docx==1.1.0      # DOCX parsing
chardet==5.2.0          # Character encoding detection

# Web UI
flask==3.0.0
werkzeug==3.0.1
flask-login==0.6.3
bcrypt>=4.0.1  # Password hashing for user authentication
gunicorn==21.2.0  # Production WSGI server
authlib==1.3.0  # OAuth 2.0 for Google and LinkedIn login

# Optional - for future enhancements
# schedule==1.2.0  # Alternative to cron
